Fixing clippy errors

+++ src/api/windows/winapi32/kernel32/sizeof_resource.rs
@@ -3,7 +3,7 @@ use crate::winapi::helper;
 
 pub fn SizeofResource(emu: &mut emu::Emu) {
     let hModule = emu.regs().rcx;
-    let hResInfo = emu.regs().rdx as u64;
+    let hResInfo = emu.regs().rdx;
 
     emu.stack_pop32(false);
     emu.stack_pop32(false);
